pyproject.toml
[project]
name = "ai-personal-brand-assistant"
version = "0.1.0"
description = "Chat assistant for your resume and personal brand"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=4.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pypdf>=4.0.0",
    "requests>=2.28.0",
    "numpy>=1.24.0",
]

[project.scripts]
personal-brand-assistant = "app:main"
